Move FunctionKind to SPI
