+++ core/trino-main/src/main/java/io/trino/metadata/GlobalFunctionCatalog.java
@@ -34,9 +34,9 @@ import java.util.function.Function;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkState;
 import static com.google.common.collect.ImmutableMap.toImmutableMap;
-import static io.trino.metadata.FunctionKind.AGGREGATE;
 import static io.trino.metadata.Signature.isOperatorName;
 import static io.trino.metadata.Signature.unmangleOperator;
+import static io.trino.spi.function.FunctionKind.AGGREGATE;
 import static io.trino.spi.type.BigintType.BIGINT;
 import static io.trino.spi.type.BooleanType.BOOLEAN;
 import static io.trino.spi.type.IntegerType.INTEGER;
