+++ core/trino-main/src/main/java/io/trino/metadata/ResolvedFunction.java
@@ -26,6 +26,7 @@ import io.airlift.json.JsonCodec;
 import io.airlift.json.JsonCodecFactory;
 import io.airlift.json.ObjectMapperProvider;
 import io.trino.collect.cache.NonEvictableLoadingCache;
+import io.trino.spi.function.FunctionKind;
 import io.trino.spi.type.Type;
 import io.trino.spi.type.TypeId;
 import io.trino.spi.type.TypeSignature;
