+++ core/trino-spi/src/main/java/io/trino/spi/function/FunctionKind.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.metadata;
+package io.trino.spi.function;
 
 public enum FunctionKind
 {
